fix: sign in api 修正
