+++ next-app/Api/Auth/SignInApi.ts
@@ -1,9 +1,26 @@
-import { SignInReqData, SignInApiResData, SignInResData } from "./Models/SignInApiModel"
-
 const baseUrl = 'http://localhost:3001/'
 
+export interface SignInReqData {
+  email: string;
+  password: string;
+}
+
+export interface SignInResData {
+  data: {
+    id: string;
+    email: string;
+    provider: string;
+    uid: string;
+    allow_password_change: boolean;
+    name: string;
+    role: number;
+    created_at: string;
+    updated_at: string;
+  }
+}
+
 export const SignInApi= async (signInReqData: SignInReqData)=> {
-  const res: Response = await fetch(baseUrl+'auth/sign_in/', {
+  return fetch(baseUrl+'auth/sign_in/', {
     method: 'POST',
     mode: 'cors',
     headers: {
@@ -15,23 +32,19 @@ export const SignInApi= async (signInReqData: SignInReqData)=> {
       password: signInReqData.password,
     })
   })
-  
-  localStorage.setItem("uid", res.headers.get("uid") ?? "")
-  localStorage.setItem("access-token", res.headers.get("access-token") ?? "")
-  localStorage.setItem("client", res.headers.get("client") ?? "")
-  
-  if(res.ok){
-    const signInResData = await res.json() as SignInResData
-    const signInApiResData: SignInApiResData = {
-      ok: true,
-      res: signInResData,
+  .then(response => {
+    if(!response.ok){
+      throw new Error("エラー発生");
     }
-    return signInApiResData
-  }else{
-    const signInApiResData: SignInApiResData = {
-      ok: false,
-      errorText: res.statusText,
-    }
-    return signInApiResData
-  }
+    localStorage.setItem("uid", response.headers.get("uid") ?? "")
+    localStorage.setItem("access-token", response.headers.get("access-token") ?? "")
+    localStorage.setItem("client", response.headers.get("client") ?? "")
+    return response.json()
+  })
+  .then(data=>{
+    return data as SignInResData
+  })
+  .catch(error=> {
+    throw new Error("エラー発生");
+  });
 }
