+++ next-app/src/pages/sign_in/index.tsx
@@ -1,7 +1,6 @@
 import type { NextPage } from 'next'
 import Link from 'next/link'
-import { SignInReqData, SignInApiResData } from 'Api/Auth/Models/SignInApiModel'
-import { SignInApi } from 'Api/Auth/SignInApi'
+import { SignInApi, SignInReqData, SignInResData } from 'Api/Auth/SignInApi'
 import { useRouter } from 'next/router'
 import { Row, Col, Card, Form, Button, FloatingLabel } from "react-bootstrap"
 
@@ -15,14 +14,14 @@ const SignInPage: NextPage = () => {
       email: email.value,
       password: password.value,
     }
-    const signInApiResData: SignInApiResData  = await SignInApi(signInReqData)
-    if(signInApiResData.ok){
+    try{
+      const signInResData: SignInResData  = await SignInApi(signInReqData)
       //User情報を保存して/homeにリダイレクト
       //router.push('/home')
-      console.log("Success SignInApi:",signInApiResData)
-    }else{
+      console.log("Success SignInApi:",signInResData)
+    }catch(e){
       //error情報を表示
-      console.log("SignInError:", signInApiResData.errorText)
+      console.error(e)
     }
   }
   return(
